Organize TheMap logic
